refactor: move project_id -> project_number translation to gcloud_context
