+++ src/xpk/core/cluster.py
@@ -15,8 +15,6 @@ limitations under the License.
 """
 
 import yaml
-from google.api_core.exceptions import PermissionDenied
-from google.cloud import resourcemanager_v3
 from kubernetes import client as k8s_client
 from kubernetes import config
 from kubernetes.client.exceptions import ApiException
@@ -32,6 +30,7 @@ from .commands import (
 from .gcloud_context import (
     add_zone_and_project,
     get_cluster_location,
+    get_project_number,
     zone_to_region,
 )
 from .nodepool import recreate_nodes_in_existing_node_pools
@@ -389,33 +388,10 @@ def get_all_clusters_programmatic(args) -> tuple[list[str], int]:
   return raw_cluster_output.splitlines(), 0
 
 
-def project_id_to_project_number(project_id: str) -> str:
-  client = resourcemanager_v3.ProjectsClient()
-  request = resourcemanager_v3.GetProjectRequest()
-  request.name = f'projects/{project_id}'
-  try:
-    response = client.get_project(request=request)
-  except PermissionDenied as e:
-    xpk_print(
-        f"Couldn't translate project id: {project_id} to project number."
-        f' Error: {e}'
-    )
-    xpk_exit(1)
-  parts = response.name.split('/', 1)
-  xpk_print(f'Project number for project: {project_id} is {parts[1]}')
-  return str(parts[1])
-
-
 def setup_k8s_env(args) -> k8s_client.ApiClient:
   add_zone_and_project(args)
   get_cluster_credentials(args)
-  # Use provided project number if available, otherwise fetch via API
-  if getattr(args, 'project_number', None):
-    xpk_print(f'Using provided project number: {args.project_number}')
-  elif args.dry_run:
-    args.project_number = abs(hash(args.project) % (10**12))  # 12 digit hash
-  else:
-    args.project_number = project_id_to_project_number(args.project)
+  args.project_number = get_project_number(args)
 
   config.load_kube_config()
   return k8s_client.ApiClient()
